Address review: move close_socket to sock.c, bump drain timeout to 10s
